refactor(backend): isolate deployment chat proxy
