+++ shared/tsconfig.json
@@ -10,8 +10,15 @@
     "strict": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
+    "forceConsistentCasingInFileNames": true,
+    "ignoreDeprecations": "6.0"
   },
-  "include": ["types/**/*", "api/**/*"],
-  "exclude": ["node_modules", "dist"]
+  "include": [
+    "types/**/*",
+    "api/**/*"
+  ],
+  "exclude": [
+    "node_modules",
+    "dist"
+  ]
 }
